requirements/local.txt
-r base.txt

Werkzeug==0.14.1  # pyup: < 0.15 # https://github.com/pallets/werkzeug
ipdb==0.12  # https://github.com/gotcha/ipdb
Sphinx==2.1.2  # https://github.com/sphinx-doc/sphinx
psycopg2==2.8.3 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
tox==3.13.2
mypy==0.711  # https://github.com/python/mypy
pytest==5.0.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar
coverage==4.5.3
django-coverage==1.2.4
django-coverage-plugin==1.6.0

# Code quality
# ------------------------------------------------------------------------------
flake8==3.7.7  # https://github.com/PyCQA/flake8
coverage==4.5.3  # https://github.com/nedbat/coveragepy
black==19.3b0  # https://github.com/ambv/black
pylint-django==2.0.10  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery

# Django
# ------------------------------------------------------------------------------
factory-boy==2.12.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==2.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.1.9  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.6.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.5.1  # https://github.com/pytest-dev/pytest-django
